reintroduced long long

+++ Avisynth3/trunk/avisynth/src/clip/onechild/videoprocess.h
@@ -1,4 +1,4 @@
-// Avisynth v3.0 alpha.  Copyright 2004 Ben Rudiak-Gould et al.
+// Avisynth v3.0 alpha.  Copyright 2004 David Pierre - Ben Rudiak-Gould et al.
 // http://www.avisynth.org
 
 // This program is free software; you can redistribute it and/or modify
@@ -47,7 +47,7 @@ class VideoProcess : public virtual OneChild
 
 public:  //clip general interface
 
-  virtual void GetAudio(void * buffer, int64 start, int count) const { return GetChild()->GetAudio(buffer, start, count); }
+  virtual void GetAudio(void * buffer, long long start, int count) const { return GetChild()->GetAudio(buffer, start, count); }
 
 
 private:  //Refactor methods
